+++ CONY-BE/payment/src/main/java/com/cony/payment/domain/purchase/service/PurchaseService.java
@@ -46,14 +46,14 @@ public class PurchaseService {
 
         validatePurchase(buyer, sale);
 
-        Long purchasePrice = sale.getSalePrice();
+        Integer purchasePrice = sale.getSalePrice();
 
         // 포인트 이동 (구매자 차감 / 판매자 지급)
-        buyer.deductPoint(purchasePrice);
+        buyer.deductPoint(Long.valueOf(purchasePrice));
 
         User seller = userRepository.findById(sale.getSellerId())
                 .orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
-        seller.chargePoint(purchasePrice);
+        seller.chargePoint(Long.valueOf(purchasePrice));
 
         sale.soldOut();
 
@@ -65,8 +65,8 @@ public class PurchaseService {
         purchaseRepository.save(purchase);
 
         // 거래 로그 기록
-        transactionService.createTransaction(buyer, TransactionType.PURCHASE, -purchasePrice, "기프티콘 구매");
-        transactionService.createTransaction(seller, TransactionType.SALE, purchasePrice, "기프티콘 판매 수익");
+        transactionService.createTransaction(buyer, TransactionType.PURCHASE, -Long.valueOf(purchasePrice), "기프티콘 구매");
+        transactionService.createTransaction(seller, TransactionType.SALE, Long.valueOf(purchasePrice), "기프티콘 판매 수익");
 
         log.info("구매 거래 완료: purchaseId={}", purchase.getId());
     }
