+++ CONY-BE/payment/src/main/java/com/cony/payment/domain/sale/dto/SaleUpdateRequestDto.java
@@ -13,5 +13,5 @@ public class SaleUpdateRequestDto {
 
     @NotNull(message = "판매 가격은 필수입니다.")
     @Min(value = 100, message = "판매 가격은 최소 100원 이상이어야 합니다.")
-    private Long salePrice;
+    private Integer salePrice;
 }
